Move HasUncertainty definition to Uncertainty.hs
